Update cli to use typer, add generate-keys command, move app import
This removes the previous cli which used argparse in favor of typer.
It allows us to more sanely create cli commands without large amount
of boilerplate or the large amount of decorators usually used by click
to convey type information. This also adds the generate-keys command
back to the cli which had been removed in previous versions as we
went through a couple of iterations on how we wanted to encrypt data
sent to skills. The SkillApplication import has now been removed from
__init__.py as well as since it is based on mindmeld it takes several
seconds to start and some dependencies such as numpy log large numbers
of warnings by default both of which make for an unpleasant CLI
experience.