Initial pass of SDK version 2 (#28)
Initial pass of SDK version 2

This represents a substantial rewrite of the existing SDK. The broad brush strokes are that it replaces the flask based
API with one which uses FastAPI, provides library methods for cryptography operations, adds a subset of the 
functionality provided by mindmeld and attempts to offer a better getting started experience. There have also been
significant updates to the CLI to add commands for easier setup and testing. This is still a work in progress and
things may be incomplete or inaccurate.

Co-authored-by: snow0x2d0 <james@snow.engineer>
Co-authored-by: Juan Rodriguez <angel.neb@gmail.com>