Raise exceptions on failed signature verification or decryption. (#30)
* Raise exceptions on failed signature verification or decryption.

The methods of our middleware that act as filters on the "receive" event
(verify signature and receive_decrypted) now raise on failures. This is essentially a
temporary hack around some behavior in starlette's router which creates a new Request
object, but using the same `receive` method passed to the middlewares, prior to
executing a route handler. As a result if the request's body is read prior to being used in that
new Request object any attempt to access the request's body will hang forever. If everything goes
well it's not a problem as the messages come through as expected. If a signature is invalid or
a message can't be decrypted however we now simply raise. This is somewhat less
than ideal as it relies on the underlying behavior of FastAPI to catch the raised exception
and return a 400 response. This should definitely be replaced with a more robust solution when time permits.