Add a check endpoint to the base API, add a CLI method to call it. (#32)
* Add a check endpoint to the base API, add a CLI method to call it.

Adds a GET endpoint /check to the base api, so that we can check if the service is up and configured
correctly. The CLI command functions similarly to invoke but can't use the configured url since it refers to the
parse endpoint. As an enhancement later we should look at perhaps adding a second check url to the configuration.