Fixes a couple dialogue manager bugs
The mindmeld API was missing support for a default handler, that argument has now been added. Additionally, the
dialogue manager was originally intended to pass the entities into queries which used entities, which was missed.
Instead, now if the handler specifies a second argument, the entire query object (either a processed query or a string)
is passed as a second argument.