V2 (#40)
SDK version 2

This represents a substantial rewrite of the existing SDK. The broad brush strokes are that it replaces the flask based
API with one which uses FastAPI, provides library methods for cryptography operations, adds a subset of the 
functionality provided by mindmeld and attempts to offer a better getting started experience. There have also been
significant updates to the CLI to add commands for easier setup and testing.  We're keeping the repository the same (name change pending) but will be publishing to a new package name (webex-skills) and starting with version 2.0 to avoid any tag conflicts in github. The previous version will be maintained on a v1 branch for a short period while we allow beta users time to transition to the new version.

Co-authored-by: snow0x2d0 <james@snow.engineer>
Co-authored-by: Juan Rodriguez <angel.neb@gmail.com>