Add uvicorn as a dependency, cleanup imports. (#41)
uvicorn was originally listed as a dev dependency, however given its usefulness in the initial developer experience we've
decided to make it a proper dependency rather than an extra. We're trying to strike a balance between developer convenience
and package bloat and for the time being at least we've decided to lean toward the former.