Split get_skill_config into more specific functions. (#53)
* Split get_skill_config into more specific functions.

The get_skill_config method previously attempted to handle a number of cases such as when config
doesn't exist, getting a config for specific skill, or getting all skill configurations. This resulted
in the introduction of a bug in PR #50 where the init command attempted to determine if a skill with the
passed in name already existed. The get_skill_config method would make an empty default configuration
if one didn't exist, but seeing a name passed in would then attempt to read from that empty value and
throw an error. This resulted in a new user being unable to initialize a project. To hopefully avoid
other issues in the future the 4 functions previously contained in the get_skill_config method have
been split into their own smaller methods which the commands can make use of as needed.