Add support for specifying context values when invoking a skill.
This adds optional flags for org/user/device ids to the `invoke` cli command.
Previously we were passing in an empty context however it may be useful, or
even necessary for testing a skill for those ids to be present. If not specified
on the command line a random UUID will be generated.