Use a fixed context in the `invoke` cli command.
Context is immutable and is not able to be changed by the skill. The
invoke command in the CLI however expected that on turns after the first
the context would be provided by the skill response. This led to the
context working on the first request but being empty on all subsequent requests. The context is now set to a default value before calling a skill
and the requests reuse the same context value until exit.