Fix bug introduced by linting cleanup related to encrypted payloads.
Previously the challenge value of an invoke request payload was inlined into the request
body and the request's challenge field was compared against the response. This however
was not possible when the payload was encrypted and led to a KeyError trying to access it.
The challenge value has been moved into its own variable for comparison in both encrypted
and non-encrypted payloads. This is a temporary fix while we rework the cli portion
dealing with invoking/running skills to be more flexible.